feature: Do a HTTP requests without expect a response
Some times we want only to send a request to the server and we doesn't care about the resposne (success or not).
For that reason we change the type "Request" that only request without assert the result and we created a new
type "RequestExpectant" that can assert expected response from the server.